some more cleanups
